+++ windirstat/HelpersTasks.h
@@ -38,7 +38,8 @@ using HashAlgorithm = enum HashAlgorithm {
     HASH_SHA1,
     HASH_SHA256,
     HASH_SHA384,
-    HASH_SHA512
+    HASH_SHA512,
+    HASH_XXHASH
 };
 
 struct HashAlgorithmInfo
@@ -53,7 +54,8 @@ constexpr std::array HashAlgorithms = {
     HashAlgorithmInfo{ HASH_SHA1, BCRYPT_SHA1_ALGORITHM, L"SHA1" },
     HashAlgorithmInfo{ HASH_SHA256, BCRYPT_SHA256_ALGORITHM, L"SHA256" },
     HashAlgorithmInfo{ HASH_SHA384, BCRYPT_SHA384_ALGORITHM, L"SHA384" },
-    HashAlgorithmInfo{ HASH_SHA512, BCRYPT_SHA512_ALGORITHM, L"SHA512" }
+    HashAlgorithmInfo{ HASH_SHA512, BCRYPT_SHA512_ALGORITHM, L"SHA512" },
+    HashAlgorithmInfo{ HASH_XXHASH, nullptr, L"xxHash" }
 };
 
 // Used at runtime to distinguish between mount points and junction points since they